+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @baitasolucoes-ui/react
 
+## 2.1.0
+
+### Minor Changes
+
+  - Adding Toast and Tooltip components
+  - Removing _test_ color on tokens
+
 ## 2.0.0
 
 ### Major Changes